Add AJAX Script & Set Rails For Multiple Formats
Add jQuery Script AJAX function which will automatically set the data
type as a script and then send a Request to Rails. Rails will then find
out which format the user want, this is case JavaScript, and then send
over the a JavaScript to be automatically execute on the client side.