Add Scrolling Detection To Know Each Bottom
Add in an event to detected when the user scrolled and if they reach
the button of the page because once the user reaches the button the
next data should load up. Also this will load the pagination when the
document loads since there maybe need to load more pages of data then
just one. Furthermore this will also check if there is a pagination to
use.